+++ x86_64/pmio.c
@@ -6,7 +6,7 @@
 int pmio_read8(uintptr_t offset, uint8_t *restrict value)
 {
     if(offset <= PMIO_MAX_OFFSET) {
-        asm volatile("inb %1, %0":"=a"(value[0]):"Nd"(offset));
+        asm volatile("inb %1, %0":"=a"(value[0]):"Nd"((uint16_t)offset));
         return 0;
     }
 
@@ -16,7 +16,7 @@ int pmio_read8(uintptr_t offset, uint8_t *restrict value)
 int pmio_read16(uintptr_t offset, uint16_t *restrict value)
 {
     if(offset <= PMIO_MAX_OFFSET) {
-        asm volatile("inw %1, %0":"=a"(value[0]):"Nd"(offset));
+        asm volatile("inw %1, %0":"=a"(value[0]):"Nd"((uint16_t)offset));
         return 0;
     }
 
@@ -26,7 +26,7 @@ int pmio_read16(uintptr_t offset, uint16_t *restrict value)
 int pmio_read32(uintptr_t offset, uint32_t *restrict value)
 {
     if(offset <= PMIO_MAX_OFFSET) {
-        asm volatile("inl %1, %0":"=a"(value[0]):"Nd"(offset));
+        asm volatile("inl %1, %0":"=a"(value[0]):"Nd"((uint16_t)offset));
         return 0;
     }
 
@@ -36,7 +36,7 @@ int pmio_read32(uintptr_t offset, uint32_t *restrict value)
 int pmio_write8(uintptr_t offset, uint8_t value)
 {
     if(offset <= PMIO_MAX_OFFSET) {
-        asm volatile("outb %0, %1"::"a"(value), "Nd"(offset));
+        asm volatile("outb %0, %1"::"a"(value), "Nd"((uint16_t)offset));
         return 0;
     }
 
@@ -46,7 +46,7 @@ int pmio_write8(uintptr_t offset, uint8_t value)
 int pmio_write16(uintptr_t offset, uint16_t value)
 {
     if(offset <= PMIO_MAX_OFFSET) {
-        asm volatile("outw %0, %1"::"a"(value), "Nd"(offset));
+        asm volatile("outw %0, %1"::"a"(value), "Nd"((uint16_t)offset));
         return 0;
     }
 
@@ -56,7 +56,7 @@ int pmio_write16(uintptr_t offset, uint16_t value)
 int pmio_write32(uintptr_t offset, uint32_t value)
 {
     if(offset <= PMIO_MAX_OFFSET) {
-        asm volatile("outl %0, %1"::"a"(value), "Nd"(offset));
+        asm volatile("outl %0, %1"::"a"(value), "Nd"((uint16_t)offset));
         return 0;
     }
 
@@ -66,5 +66,5 @@ int pmio_write32(uintptr_t offset, uint32_t value)
 void pmio_wait(void)
 {
     // FIXME: linux has a lot more here...
-    asm volatile("outb %0, $0x80"::"a"(0));
+    asm volatile("outb %0, $0x80"::"a"((uint8_t)0));
 }
